Refactor UserManager and Replication interfaces.
