+++ pkg/controller/ps/controller.go
@@ -45,13 +45,13 @@ import (
 	apiv1alpha1 "github.com/percona/percona-server-mysql-operator/api/v1alpha1"
 	"github.com/percona/percona-server-mysql-operator/pkg/clientcmd"
 	"github.com/percona/percona-server-mysql-operator/pkg/controller/psrestore"
+	"github.com/percona/percona-server-mysql-operator/pkg/db"
 	"github.com/percona/percona-server-mysql-operator/pkg/haproxy"
 	"github.com/percona/percona-server-mysql-operator/pkg/k8s"
 	"github.com/percona/percona-server-mysql-operator/pkg/mysql"
 	"github.com/percona/percona-server-mysql-operator/pkg/mysqlsh"
 	"github.com/percona/percona-server-mysql-operator/pkg/orchestrator"
 	"github.com/percona/percona-server-mysql-operator/pkg/platform"
-	"github.com/percona/percona-server-mysql-operator/pkg/replicator"
 	"github.com/percona/percona-server-mysql-operator/pkg/router"
 	"github.com/percona/percona-server-mysql-operator/pkg/util"
 )
@@ -204,10 +204,7 @@ func (r *PerconaServerMySQLReconciler) deleteMySQLPods(ctx context.Context, cr *
 		firstPodFQDN := fmt.Sprintf("%s.%s.%s", firstPod.Name, mysql.ServiceName(cr), cr.Namespace)
 		firstPodUri := fmt.Sprintf("%s:%s@%s", apiv1alpha1.UserOperator, operatorPass, firstPodFQDN)
 
-		db, err := replicator.NewReplicator(&firstPod, r.ClientCmd, apiv1alpha1.UserOperator, operatorPass, firstPodFQDN)
-		if err != nil {
-			return errors.Wrapf(err, "connect to %s", firstPod.Name)
-		}
+		um := db.NewReplicationManager(&firstPod, r.ClientCmd, apiv1alpha1.UserOperator, operatorPass, firstPodFQDN)
 
 		mysh, err := mysqlsh.NewWithExec(r.ClientCmd, &firstPod, firstPodUri)
 		if err != nil {
@@ -222,12 +219,12 @@ func (r *PerconaServerMySQLReconciler) deleteMySQLPods(ctx context.Context, cr *
 
 			podFQDN := fmt.Sprintf("%s.%s.%s", pod.Name, mysql.ServiceName(cr), cr.Namespace)
 
-			state, err := db.GetMemberState(ctx, podFQDN)
+			state, err := um.GetMemberState(ctx, podFQDN)
 			if err != nil {
 				return errors.Wrapf(err, "get member state of %s from performance_schema", pod.Name)
 			}
 
-			if state == replicator.MemberStateOffline {
+			if state == db.MemberStateOffline {
 				log.Info("Member is not part of GR or already removed", "member", pod.Name, "memberState", state)
 				continue
 			}
@@ -1008,12 +1005,9 @@ func (r *PerconaServerMySQLReconciler) getPrimaryFromGR(ctx context.Context, cr 
 		return "", err
 	}
 
-	db, err := replicator.NewReplicator(firstPod, r.ClientCmd, apiv1alpha1.UserOperator, operatorPass, fqdn)
-	if err != nil {
-		return "", errors.Wrapf(err, "open connection to %s", fqdn)
-	}
+	um := db.NewReplicationManager(firstPod, r.ClientCmd, apiv1alpha1.UserOperator, operatorPass, fqdn)
 
-	return db.GetGroupReplicationPrimary(ctx)
+	return um.GetGroupReplicationPrimary(ctx)
 }
 
 func (r *PerconaServerMySQLReconciler) getPrimaryHost(ctx context.Context, cr *apiv1alpha1.PerconaServerMySQL) (string, error) {
@@ -1059,10 +1053,7 @@ func (r *PerconaServerMySQLReconciler) stopAsyncReplication(ctx context.Context,
 			if err != nil {
 				return err
 			}
-			repDb, err := replicator.NewReplicator(pod, r.ClientCmd, apiv1alpha1.UserOperator, operatorPass, hostname)
-			if err != nil {
-				return errors.Wrapf(err, "connect to replica %s", hostname)
-			}
+			repDb := db.NewReplicationManager(pod, r.ClientCmd, apiv1alpha1.UserOperator, operatorPass, hostname)
 
 			if err := orchestrator.StopReplicationExec(gCtx, r.ClientCmd, orcPod, hostname, port); err != nil {
 				return errors.Wrapf(err, "stop replica %s", hostname)
@@ -1073,7 +1064,7 @@ func (r *PerconaServerMySQLReconciler) stopAsyncReplication(ctx context.Context,
 				return errors.Wrapf(err, "get replication status of %s", hostname)
 			}
 
-			for status == replicator.ReplicationStatusActive {
+			for status == db.ReplicationStatusActive {
 				time.Sleep(250 * time.Millisecond)
 				status, _, err = repDb.ReplicationStatus(ctx)
 				if err != nil {
@@ -1116,13 +1107,10 @@ func (r *PerconaServerMySQLReconciler) startAsyncReplication(ctx context.Context
 			if err != nil {
 				return err
 			}
-			db, err := replicator.NewReplicator(pod, r.ClientCmd, apiv1alpha1.UserOperator, operatorPass, hostname)
-			if err != nil {
-				return errors.Wrapf(err, "get db connection to %s", hostname)
-			}
+			um := db.NewReplicationManager(pod, r.ClientCmd, apiv1alpha1.UserOperator, operatorPass, hostname)
 
 			log.V(1).Info("Change replication source", "primary", primary.Key.Hostname, "replica", hostname)
-			if err := db.ChangeReplicationSource(ctx, primary.Key.Hostname, replicaPass, primary.Key.Port); err != nil {
+			if err := um.ChangeReplicationSource(ctx, primary.Key.Hostname, replicaPass, primary.Key.Port); err != nil {
 				return errors.Wrapf(err, "change replication source on %s", hostname)
 			}
